File restructuring. Made folder for keys.
