+++ src/keys/displaydata/displayData.jsx
@@ -1,0 +1,24 @@
+import React, { Component } from 'react'
+
+class DisplayData extends Component {
+
+  constructor(props, context) {
+    super(props, context)
+    this.state = {
+      id: 0,
+      text: 'Stuff',
+      done: false
+    }
+  }
+
+  render() {
+    return (
+      <div>
+
+      </div>
+    )
+  }
+
+}
+
+export default DisplayData
